requirements/base.txt
# These are the reqs to build docs and run tests.
sphinx
django-nose
django-redis
jinja2
redis
coverage
psycopg2
pymemcache
